SQL file for congressional schema
This can be used to create the entire DB from the command line.  However, it's also a nuclear option - it drops all the tables if they exist already, necessitating that they be re-scraped.  So, you can use this if for some reason we need to repopulate the entire DB from CSVs, but don't use this if you need to repopulate a single table (just DROP TABLE manually where applicable).